added the homepage

+++ backend/controllers/postsController.js
@@ -1,4 +1,5 @@
 import Post from '../models/post';
+import post from '../models/post';
 
 class PostController {
 
@@ -12,41 +13,53 @@ class PostController {
       return res.status(200).json({
         message: 'Successfully fetched posts',
         posts: posts.map(post => {
-            const { _id, title, description, dateCreated, category } = post;
+          const { _id, title, description, dateCreated, category, creator } = post;
             return {
               id: _id,
               title,
               description,
               category,
-              dateCreated
+              dateCreated,
+              creator
             }
         })
       })
      });
   }
-  static getPostsDetail(req, res) {
-    Post.find((err, posts) => {
-      console.log(posts);
-      if (err) {
-         return res.status(500).json({
-           error: err
-         });
-      }
-     return res.status(200).json({
-       message: 'Successfully fetched posts',
-       posts: posts.map(post => {
-
-           const { _id, title, description, category, dateCreated } = post;
-           return {
-             id: _id,
-             title,
-             description,
-             category,
-             dateCreated
-           }
-       })
-     })
-    });
+  static async getPostsDetail(req, res) {
+    let { pageSize, pageNumber } = req.query;
+    if (typeof(pageSize) === 'undefined' && typeof(pageNumber) === 'undefined') {
+      pageSize = 10;
+      pageNumber = 1;
+    }
+    try {
+      console.log('ddd', typeof(pageNumber), typeof(pageSize));
+      let postData;
+      const count = await Post.countDocuments();
+      console.log('count', count);
+       Post.find({}, (err, posts) => {
+        return res.status(200).json({
+          message: 'Successfully fetched posts',
+          posts:   posts.map(post => {
+              const { _id, title, description, category, dateCreated, imagePath, creator } = post;
+              return {
+                id: _id,
+                title,
+                description,
+                category,
+                dateCreated,
+                imagePath,
+                creator
+              }
+          }),
+          totalItems: count
+        });
+      }).skip((+pageNumber - 1) * +pageSize).limit(+pageSize);
+    } catch (error) {
+      return res.status(500).json({
+        error
+      });
+    }
   }
 
   static getPost(req, res) {
@@ -61,7 +74,7 @@ class PostController {
           message: 'Post was not found'
         });
       }
-      const { _id, title, description, category, dateCreated } = post;
+      const { _id, title, description, category, dateCreated, imagePath } = post;
       return res.status(200).json({
         message: 'Post was fetched successfully',
         post: {
@@ -69,7 +82,8 @@ class PostController {
           title,
           description,
           category,
-          dateCreated
+          dateCreated,
+          imagePath
         }
       });
     });
@@ -78,10 +92,14 @@ class PostController {
 
   static async addPost(req, res) {
     const {title, description, category} = req.body;
+    const url = `${req.protocol}://${req.get('host')}/images`
     const post = new Post({
       title,
       description,
-      category
+      category,
+      imagePath: `${url}/${req.file.filename}`,
+      creator: req.userData.userId
+
     });
     try {
       const savedPost = await post.save();
@@ -92,24 +110,66 @@ class PostController {
           title,
           description,
           category,
-          dateCreated: savedPost.dateCreated
+          imagePath,
+          creator: savedPost.creator
         }
       });
     } catch (error) {
-      res.status(500).json({
+      return res.status(500).json({
         error
       });
     }
   }
+  static editPost(req, res) {
+    const {title, description, category, creator} = req.body;
+    const loggedInUser = req.userData.userId;
+    if (creator !== loggedInUser) {
+      return res.status(401).json({
+        error: 'Unauthorized access'
+      });
+    }
+    let imagePath = req.body.imagePath;
+    if(req.file) {
+      const url = `${req.protocol}://${req.get('host')}/images`
+      imagePath = `${url}/${req.file.filename}`
+    }
+
+    const updatedPost = {
+      id: req.params.id,
+      title,
+      description,
+      category,
+      imagePath
+    }
+    Post.updateOne({_id: req.params.id}, updatedPost, { new: true }, (err, result) => {
+      if (err) {
+        return res.status(500).json({
+          error: 'Error occurred while updating post'
+        });
+        }
+        console.log(result.nModified);
+        if (result.nModified > 0) {
+          return res.status(200).json({
+            message: 'Succeessully updated post',
+            imagePath
+          });
+        }
+        return res.status(404).json({
+          message: 'Post was not found'
+        });
+    });
+  }
   static deletePost(req, res) {
     const postId = req.params.id;
-    Post.deleteOne({_id: postId}, (err, result) => {
+    const loggedInUser = req.userData.userId;
+
+    Post.deleteOne({_id: postId, creator: loggedInUser}, (err, result) => {
       if (err) {
         return res.status(500).json({ error: err});
       }
       if (result.n === 0) {
-        return res.status(404).json({
-          message: 'Post was not Found'
+        return res.status(401).json({
+          message: 'Unauthorized'
         });
       }
       return res.status(200).json({
